+++ back/src/main/java/com/stock/survive/controller/MyPageController.java
@@ -24,17 +24,17 @@ public class MyPageController {
 
     // 메인/리스트 별표(★) 표시에 최적
     @GetMapping("/me/favorite-ids")
-    public Set<Integer> favoriteIds(Authentication auth) {
+    public Set<Long> favoriteIds(Authentication auth) {
         return myPageQueryService.getFavoriteIdSet(extractUid(auth));
     }
 
-    private Integer extractUid(Authentication auth) {
+    private Long extractUid(Authentication auth) {
         if (auth == null || !auth.isAuthenticated())
             throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "UNAUTHORIZED");
         Object p = auth.getPrincipal();
-        if (p instanceof Integer i) return i;
-        if (p instanceof Number n) return n.intValue();
-        try { return Integer.valueOf(String.valueOf(p)); } catch (Exception ignored) {}
+        if (p instanceof Long i) return i;
+        if (p instanceof Number n) return (long) n.intValue();
+        try { return Long.valueOf(Integer.valueOf(String.valueOf(p))); } catch (Exception ignored) {}
         throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "INVALID_PRINCIPAL");
     }
 }
